Add files via upload
Task 1: To Explore Supervised Machine Learning- Linear Regression:
In this regression task I have predicted the percentage of marks that a student is expected to score based upon the number of hours he studied. This task is based on Linear Regression. I have used Python libraries like numpy, pandas, matplotlib and scikit-learn for building the algorithm.